Restructured Rent-Data and all files that require the module.
