+++ Rent/Rent-Data/src/main/java/com/rent/data/dataaccess/api/dao/user/UserInfoDAO.java
@@ -1,6 +1,6 @@
-package dataaccess.api.dao.User;
+package com.rent.data.dataaccess.api.dao.user;
 
-import dataaccess.api.entities.user.UserInfo;
+import com.rent.data.dataaccess.api.entities.user.UserInfo;
 
 /**
  * Created by Duck on 11/11/2016.
