+++ Rent/Rent-Data/src/main/java/com/rent/data/dataaccess/api/entities/listing/PriceCategory.java
@@ -1,4 +1,4 @@
-package dataaccess.api.entities;
+package com.rent.data.dataaccess.api.entities.listing;
 
 import javax.persistence.*;
