+++ Rent/Rent-Data/src/main/java/com/rent/data/dataaccess/auth/dao/registration/RegistrationTokenDAOImpl.java
@@ -2,7 +2,6 @@ package com.rent.data.dataaccess.auth.dao.registration;
 
 import com.rent.data.dataaccess.auth.dao.AuthDAO;
 import com.rent.data.dataaccess.auth.entities.registration.RegistrationToken;
-import dataaccess.api.dao.ApiDAO;
 import org.hibernate.Criteria;
 import org.hibernate.criterion.Restrictions;
 import org.springframework.stereotype.Repository;
