--- Rent/Rent-Data/src/main/java/dataaccess/api/dao/Registration/VerificationTokenDAO.java
@@ -1,15 +1,0 @@
-package dataaccess.api.dao.Registration;
-
-import dataaccess.api.entities.Registration.VerificationToken;
-
-/**
- * Created by Duck on 10/28/2016.
- */
-public interface VerificationTokenDAO {
-
-    boolean saveVerificationToken(VerificationToken token);
-
-    VerificationToken getTokenByString(String token);
-
-    boolean removeToken(VerificationToken verificationToken);
-}
