+++ Rent/Rent-Main/src/main/java/Controllers/UserPost/PriceCategoryController.java
@@ -1,7 +1,7 @@
 package Controllers.UserPost;
 
 import Services.UserPost.PriceCategoryService;
-import dataaccess.api.entities.PriceCategory;
+import com.rent.data.dataaccess.api.entities.listing.PriceCategory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.web.bind.annotation.*;
@@ -14,7 +14,6 @@ import java.util.List;
 
 @RestController
 @RequestMapping("/priceCategories")
-@ComponentScan("Controllers")
 public class PriceCategoryController {
 
     @Autowired
