+++ Rent/Rent-Main/src/main/java/Controllers/UserPost/UserPostController.java
@@ -3,7 +3,7 @@ package Controllers.UserPost;
 import DTOEntities.UserPost.UserPostDTO;
 import Services.UserPost.UserPostService;
 import Utilities.Filters.UserPostFilter;
-import dataaccess.api.entities.UserPost;
+import com.rent.data.dataaccess.api.entities.listing.UserPost;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.web.bind.annotation.*;
@@ -17,7 +17,6 @@ import java.util.List;
 
 @RestController
 @RequestMapping("/userPosts")
-@ComponentScan("Controllers")
 public class UserPostController {
 
     @Autowired
