+++ Rent/Rent-Main/src/main/java/DTOEntities/User/UserDTO.java
@@ -14,7 +14,7 @@ public class UserDTO implements java.io.Serializable {
 
     @NotNull(message = "username required")
     @Pattern(regexp = "^([a-zA-Z0-9_\\-\\.]+)@([a-zA-Z0-9_\\-\\.]+)\\.([a-zA-Z]{2,5})$",
-            message = "User name must be a valid email address.")
+            message = "user name must be a valid email address.")
     private String username;
 
     @NotNull(message = "password required")
