+++ Rent/Rent-Main/src/main/java/Services/UserPost/PriceCategoryService.java
@@ -1,6 +1,6 @@
 package Services.UserPost;
 
-import dataaccess.api.entities.PriceCategory;
+import com.rent.data.dataaccess.api.entities.listing.PriceCategory;
 
 import java.util.List;
